requirements.txt
aiogram==3.13.0
asyncpg
sqlalchemy[asyncio]>=2.0
aiohttp_socks
alembic                  # для миграций
python-dotenv
Pillow                   # если нужны изображения